useLocalState hook implemented + new components created for hover divs + deleted unwanted code
