+++ src/components/NavBar/NavBar.styles.jsx
@@ -294,13 +294,6 @@ export const StyledListItemSubNav = styled.li`
   list-style-type: none;
   cursor: pointer;
 `;
-export const BulletPoint = styled.div`
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background: ${(props) => props.theme.bulletPointBackground};
-  margin-right: 10px;
-`;
 export const ListItemWithProgress = styled.li`
   display: flex;
   align-items: center;
@@ -308,6 +301,14 @@ export const ListItemWithProgress = styled.li`
   list-style-type: disc;
   cursor: pointer;
 `;
+export const BulletPoint = styled.div`
+  width: 10px;
+  height: 10px;
+  border-radius: 50%;
+  background: ${(props) => props.theme.bulletPointBackground};
+  margin-right: 10px;
+`;
+
 export const ArrowTrendingDown = styled.svg`
   width: 25px;
   color: red;
@@ -354,348 +355,5 @@ export const TrendingUp = () => {
         d="M2.25 18L9 11.25l4.306 4.307a11.95 11.95 0 015.814-5.519l2.74-1.22m0 0l-5.94-2.28m5.94 2.28l-2.28 5.941"
       />
     </ArrowTrendingUp>
-  );
-};
-export const BitcoinLogoSubNav = styled.svg`
-  width: 20px;
-  margin-right: 8px;
-`;
-export const BitcoinSvgSubNav = () => {
-  return (
-    <BitcoinLogoSubNav
-      enable-background="new 0 0 2499.6 2500"
-      viewBox="0 0 2499.6 2500"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="m2499.6 1250c0 690.2-559.6 1249.8-1250.1 1249.9-690 0-1249.6-559.7-1249.5-1250-.2-690.3 559.4-1249.9 1249.7-1249.9s1249.9 559.7 1249.9 1250z"
-        fill="#eab300"
-      />
-      <g fill="#fff">
-        <path d="m1741.5 943.8c-16.1-167.4-160.6-223.5-343.2-239.5v-232.3h-141.3v226.1c-37.1 0-75.1.7-112.8 1.5v-227.6h-141.3l-.1 232.1c-30.6.6-60.7 1.2-90 1.2v-.7l-194.9-.1v151s104.4-2 102.6-.1c57.3 0 75.9 33.2 81.3 61.9v264.6c4 0 9.1.2 14.9 1h-14.9l-.1 370.7c-2.5 18-13.1 46.7-53.1 46.8 1.8 1.6-102.7 0-102.7 0l-28.1 168.8h184c34.2 0 67.9.6 100.9.8l.1 234.9h141.2v-232.4c38.7.8 76.2 1.1 112.9 1.1l-.1 231.3h141.3v-234.4c237.6-13.6 404.1-73.5 424.7-296.7 16.7-179.7-67.8-260-202.7-292.4 82.1-41.6 133.4-115.1 121.4-237.6zm-197.8 502.2c0 175.5-300.5 155.6-396.4 155.6v-311.3c95.9.2 396.4-27.3 396.4 155.7zm-65.8-439.1c0 159.7-250.8 141-330.6 141.1v-282.2c79.9 0 330.7-25.4 330.6 141.1z" />
-        <path d="m902 1175.7h21v15.5h-21z" />
-      </g>
-    </BitcoinLogoSubNav>
-  );
-};
-export const EthereumLogoSubNav = styled.svg`
-  width: 16px;
-  margin-right: 8px;
-`;
-export const EthereumSvgSubNav = () => {
-  return (
-    <EthereumLogoSubNav
-      viewBox="420.1 80.7 1079.8 1758.6"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="m959.8 80.7-539.7 895.6 539.7-245.3z" fill="#8a92b2" />
-      <path
-        d="m959.8 731-539.7 245.3 539.7 319.1zm539.8 245.3-539.8-895.6v650.3z"
-        fill="#62688f"
-      />
-      <path d="m959.8 1295.4 539.8-319.1-539.8-245.3z" fill="#454a75" />
-      <path d="m420.1 1078.7 539.7 760.6v-441.7z" fill="#8a92b2" />
-      <path d="m959.8 1397.6v441.7l540.1-760.6z" fill="#62688f" />
-    </EthereumLogoSubNav>
-  );
-};
-export const ProgressBar = styled.div`
-  width: 60px;
-  height: 14px;
-  border-radius: 20px;
-  margin-left: 8px;
-  background: #2172e5;
-`;
-export const TotalVolumeProgressBar = styled.div`
-  height: 100%;
-  width: ${(props) => props.percent.totalVolumePercent}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-  margin-left: 2px;
-`;
-const ProgressBarHover = styled.div`
-  width: 100%;
-  height: 14px;
-  border-radius: 20px;
-  background: #2172e5;
-`;
-const TotalVolumeProgressBarHover = styled.div`
-  height: 100%;
-  width: ${(props) => props.width}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-  margin-left: 2px;
-`;
-export const BtcDominanceProgressBar = styled.div`
-  height: 100%;
-  width: ${(props) => props.percent.btcDominancePercent}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-`;
-export const EthDominanceProgeressBar = styled.div`
-  height: 100%;
-  width: ${(props) => props.percent.ethDominancePercent}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-`;
-const BtcDominanceProgressBarHover = styled.div`
-  height: 100%;
-  width: ${(props) => props.width}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-`;
-const EthDominanceProgressBarHover = styled.div`
-  height: 100%;
-  width: ${(props) => props.width}%;
-  border-radius: 20px;
-  background: ${(props) => props.theme.progressBarsBackground};
-`;
-export const HoverDivsContainer = styled.div`
-  display: flex;
-`;
-const HoverDiv1 = styled.div`
-  width: 400px;
-  padding: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-left: 250px;
-  margin-top: -15px;
-`;
-const HoverDiv2 = styled.div`
-  width: 440px;
-  padding: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 400px;
-`;
-const HoverDiv3 = styled.div`
-  width: 350px;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 550px;
-`;
-const HoverDiv4 = styled.div`
-  width: 350px;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 700px;
-`;
-const HoverDiv5 = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 850px;
-`;
-const HoverDiv6 = styled.div`
-  align-items: center;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 950px;
-`;
-const HoverDiv7 = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 1080px;
-`;
-const HoverDiv8 = styled.div`
-  align-items: center;
-  padding: 10px;
-  background: ${(props) => props.theme.main};
-  border-radius: 10px;
-  z-index: 30;
-  position: absolute;
-  margin-top: -15px;
-  margin-left: 1130px;
-`;
-const StyledSpan = styled.span`
-  font-weigth: bold;
-  color: ${(props) => props.color};
-  margin-left: 8px;
-`;
-const HoverDivInner = styled.div`
-  display: flex;
-  align-items: center;
-`;
-export const CoinsHoverDiv = ({ coins }) => {
-  return (
-    <HoverDiv1>
-      <p>
-        Number of existing coins in the markget:{" "}
-        <span style={{ color: "#00FC2A", fontWeight: "bold" }}>{coins}</span>
-      </p>
-    </HoverDiv1>
-  );
-};
-export const ExchangesHoverDiv = ({ exchanges }) => {
-  return (
-    <HoverDiv2>
-      <p>
-        Number of existing exchange markets to trade at:{" "}
-        <span style={{ color: "gold", fontWeight: "bold" }}>{exchanges}</span>
-      </p>
-    </HoverDiv2>
-  );
-};
-export const TotalMarketCapHoverDiv = ({
-  totalMarketCap,
-  changePercentage,
-  color,
-}) => {
-  return (
-    <HoverDiv3>
-      <HoverDivInner>
-        <BulletPoint />
-        Total Market Cap:{" "}
-        <span style={{ fontWeight: "bold", marginLeft: "8px" }}>
-          {totalMarketCap}
-        </span>
-      </HoverDivInner>
-      <HoverDivInner>
-        <BulletPoint />
-        24h Change: <StyledSpan color={color}>{changePercentage}%</StyledSpan>
-        {changePercentage > 0 ? <TrendingUp /> : <TrendingDown />}
-      </HoverDivInner>
-    </HoverDiv3>
-  );
-};
-export const VolumeVsMarketCapDiv = ({
-  volume,
-  totalMarketCap,
-  totalVolumePercent,
-}) => {
-  return (
-    <HoverDiv4>
-      <p>24h Volume vs Total Market Cap</p>
-      <HoverDivInner>
-        <BulletPoint />
-        24h Volume:{" "}
-        <span style={{ fontWeight: "bold", marginLeft: "8px" }}>{volume}</span>
-      </HoverDivInner>
-      <HoverDivInner>
-        <BulletPoint />
-        Total Market Cap:
-        <span style={{ fontWeight: "bold", marginLeft: "8px" }}>
-          {totalMarketCap}
-        </span>
-      </HoverDivInner>
-      <ProgressBarHover>
-        <TotalVolumeProgressBarHover width={totalVolumePercent} />
-      </ProgressBarHover>
-      <p style={{ fontWeight: "bold" }}>{totalVolumePercent}%</p>
-    </HoverDiv4>
-  );
-};
-export const BtcDominanceHoverDiv1 = ({ btcDominance }) => {
-  return (
-    <HoverDiv5>
-      <BulletPoint />
-      <p>
-        Bitcoin(BTC) Dominance:{" "}
-        <span style={{ fontWeight: "bold" }}>{btcDominance}%</span>
-      </p>
-    </HoverDiv5>
-  );
-};
-export const BtcDominanceHoverDiv2 = ({
-  btcMarketCap,
-  totalMarketCap,
-  btcDominancePercent,
-}) => {
-  return (
-    <HoverDiv6>
-      BTC Market Cap vs Total Market Cap
-      <HoverDivInner>
-        <BulletPoint />
-        <p>
-          BTC Market Cap:{" "}
-          <span style={{ fontWeight: "bold" }}>{btcMarketCap}</span>
-        </p>
-      </HoverDivInner>
-      <HoverDivInner>
-        <BulletPoint />
-        <p>
-          Total Market Cap:{" "}
-          <span style={{ fontWeight: "bold" }}>{totalMarketCap}</span>
-        </p>
-      </HoverDivInner>
-      <ProgressBarHover>
-        <BtcDominanceProgressBarHover width={btcDominancePercent} />
-      </ProgressBarHover>
-      <p style={{ fontWeight: "bold" }}>{btcDominancePercent}%</p>
-    </HoverDiv6>
-  );
-};
-export const EthDominanceHoverDiv1 = ({ ethDominance }) => {
-  return (
-    <HoverDiv7>
-      <BulletPoint />
-      <p>
-        Ethereum(ETH) Dominance:{" "}
-        <span style={{ fontWeight: "bold" }}>{ethDominance}%</span>
-      </p>
-    </HoverDiv7>
-  );
-};
-export const EthDominanceHoverDiv2 = ({
-  ethMarketCap,
-  totalMarketCap,
-  ethDominancePercent,
-}) => {
-  return (
-    <HoverDiv8>
-      ETH Market Cap vs Total Market Cap
-      <HoverDivInner>
-        <BulletPoint />
-        <p>
-          ETC Market Cap:{" "}
-          <span style={{ fontWeight: "bold" }}>{ethMarketCap}</span>
-        </p>
-      </HoverDivInner>
-      <HoverDivInner>
-        <BulletPoint />
-        <p>
-          Total Market Cap:{" "}
-          <span style={{ fontWeight: "bold" }}>{totalMarketCap}</span>
-        </p>
-      </HoverDivInner>
-      <ProgressBarHover>
-        <EthDominanceProgressBarHover width={ethDominancePercent} />
-      </ProgressBarHover>
-      <p style={{ fontWeight: "bold" }}>{ethDominancePercent}%</p>
-    </HoverDiv8>
   );
 };
