+++ src/components/SubNavVolumeVsMarketCap/index.jsx
@@ -1,0 +1,40 @@
+import React from "react";
+import { useState } from "react";
+import {
+  ListItemWithProgress,
+  ProgressBar,
+  TotalVolumeProgressBar,
+  VolumeVsMarketCapDiv,
+} from "./SubNavVolumeVsMarketCap.styles";
+import { BulletPoint } from "../NavBar/NavBar.styles";
+
+export const SubNavVolumeVsMarketCap = ({
+  totalVolume,
+  text,
+  percent,
+  totalMarketCap,
+}) => {
+  const [hovered, setHovered] = useState(false);
+  return (
+    <>
+      <ListItemWithProgress
+        onMouseEnter={() => setHovered(true)}
+        onMouseLeave={() => setHovered(false)}
+      >
+        <BulletPoint />
+        {totalVolume}
+        {text}
+        <ProgressBar>
+          <TotalVolumeProgressBar percent={percent} />
+        </ProgressBar>
+      </ListItemWithProgress>
+      {hovered && (
+        <VolumeVsMarketCapDiv
+          volume={totalVolume}
+          totalMarketCap={totalMarketCap}
+          totalVolumePercent={percent}
+        />
+      )}
+    </>
+  );
+};
