+++ src/contexts/CurrencyContext.jsx
@@ -1,2 +1,2 @@
-import { createContext, useState } from "react";
+import { createContext } from "react";
 export const CurrencyContext = createContext();
