+++ src/pages/Coins/Coin/index.jsx
@@ -67,121 +67,105 @@ export const Coin = (props) => {
       case ACTIONS.GET_COIN_DATA:
         const data = action.payload;
         return {
-          coinData: data,
-          coinName: data.name,
-          coinSymbol: data.symbol.toUpperCase(),
-          coinImgSrc: data.image.large,
-          coinHomePage: data.links.homepage[0],
-          coinCurrentPrice: data.market_data.current_price[currency],
-          priceChangePercentage24h:
-            data.market_data.price_change_percentage_24h.toFixed(2),
-          coinAth: data.market_data.ath[currency],
-          coinAthDate: data.market_data.ath_date[currency].slice(0, 10),
-          coinAtl: data.market_data.atl[currency],
-          coinAtlDate: data.market_data.atl_date[currency].slice(0, 10),
-          coinMarketCap: data.market_data.market_cap[currency],
-          coinFullyDillutedValuation:
-            data.market_data.fully_diluted_valuation[currency],
-          coinVolume24h: data.market_data.total_volume[currency],
-          coinVolumeOverMarketCap: (
-            data.market_data.total_volume[currency] /
-            data.market_data.market_cap[currency]
-          ).toFixed(2),
-          coinCirculatingSupply: data.market_data.circulating_supply,
-          coinTotalSupply: data.market_data.total_supply,
-          num1:
-            (
+          coinData: {
+            ...state.coinData,
+            coinName: data.name,
+            coinSymbol: data.symbol.toUpperCase(),
+            coinImgSrc: data.image.large,
+            coinHomePage: data.links.homepage[0],
+            coinCurrentPrice: data.market_data.current_price[currency],
+            priceChangePercentage24h:
+              data.market_data.price_change_percentage_24h.toFixed(2),
+            coinAth: data.market_data.ath[currency],
+            coinAthDate: data.market_data.ath_date[currency].slice(0, 10),
+            coinAtl: data.market_data.atl[currency],
+            coinAtlDate: data.market_data.atl_date[currency].slice(0, 10),
+            coinMarketCap: data.market_data.market_cap[currency],
+            coinFullyDillutedValuation:
+              data.market_data.fully_diluted_valuation[currency],
+            coinVolume24h: data.market_data.total_volume[currency],
+            coinVolumeOverMarketCap: (
               data.market_data.total_volume[currency] /
               data.market_data.market_cap[currency]
-            ).toFixed(2) * 100,
-          num2:
-            100 -
-            (
-              data.market_data.total_volume[currency] /
-              data.market_data.market_cap[currency]
-            ).toFixed(2) *
-              100,
-          coinDescription: data.description["en"],
-          coinBlockChainSite1: data.links.blockchain_site[0],
-          coinBlockChainSite2: data.links.blockchain_site[1],
-          coinBlockChainSite3: data.links.blockchain_site[2],
-          coinPrices: data.prices.map((el) => el[1]),
-          chartLabels: data.prices.map((el) => formatDate(el[0])),
+            ).toFixed(2),
+            coinCirculatingSupply: data.market_data.circulating_supply,
+            coinTotalSupply: data.market_data.total_supply,
+            num1:
+              (
+                data.market_data.total_volume[currency] /
+                data.market_data.market_cap[currency]
+              ).toFixed(2) * 100,
+            num2:
+              100 -
+              (
+                data.market_data.total_volume[currency] /
+                data.market_data.market_cap[currency]
+              ).toFixed(2) *
+                100,
+            coinDescription: data.description["en"],
+            coinBlockChainSite1: data.links.blockchain_site[0],
+            coinBlockChainSite2: data.links.blockchain_site[1],
+            coinBlockChainSite3: data.links.blockchain_site[2],
+            coinPrices: data.prices.map((el) => el[1]),
+            chartLabels: data.prices.map((el) => formatDate(el[0])),
+          },
         };
       default:
         return state;
     }
   };
-
   const [state, dispatch] = useReducer(reducer, {
-    coinData: {},
-    coinName: "",
-    coinSymbol: "",
-    coinImgSrc: "",
-    coinHomePage: "",
-    coinCurrentPrice: 0,
-    priceChangePercentage24h: 0,
-    coinAth: 0,
-    coinAthDate: "",
-    coinAtl: 0,
-    coinAtlDate: "",
-    coinMarketCap: 0,
-    coinFullyDillutedValuation: 0,
-    coinVolume24h: 0,
-    coinVolumeOverMarketCap: 0,
-    coinCirculatingSupply: 0,
-    coinTotalSupply: 0,
-    color1: getRandomColor(),
-    color2: getRandomColor(),
-    num1: 0,
-    num2: 0,
-    coinDescription: "",
-    coinBlockChainSite1: "",
-    coinBlockChainSite2: "",
-    coinBlockChainSite3: "",
-    days: 1,
-    interval: "hourly",
-    chartLabels: [],
-    coinPrices: [],
-    isChartToday: true,
-    isChartWeek: false,
-    isChart1Month: false,
-    isChart3Months: false,
-    isChart1Year: false,
-    isChartAllTime: false,
+    coinData: {
+      days: 1,
+      interval: "hourly",
+      isChartToday: true,
+      isChartWeek: false,
+      isChart1Month: false,
+      isChart3Months: false,
+      isChart1Year: false,
+      isChartAllTime: false,
+      color1: getRandomColor(),
+      color2: getRandomColor(),
+      chartLabels: [],
+      coinPrices: [],
+    },
   });
-
   const getThemeColors = () => {
     const theme = localStorage.getItem("theme");
     return JSON.parse(theme);
   };
-
   useEffect(() => {
-    getCoinData(props.match.params.coinId, state.days, state.interval);
+    getCoinData(
+      props.match.params.coinId,
+      state.coinData.days,
+      state.coinData.interval
+    );
   }, [currency]);
 
-  localStorage.setItem("coinCurrentPrice", state.coinCurrentPrice);
+  localStorage.setItem("coinCurrentPrice", state.coinData.coinCurrentPrice);
   const theme = getThemeColors();
   const coinPricesData = {
-    labels: state.chartLabels,
+    labels: state.coinData.chartLabels,
     datasets: [
       {
         label: `${props.match.params.coinId} Price`,
-        data: state.coinPrices,
+        data: state.coinData.coinPrices,
         borderColor:
-          state.coinPrices[0] < state.coinPrices[state.coinPrices.length - 1]
+          state.coinData.coinPrices[0] <
+          state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
             ? theme.btcPriceChartBorderColorGain
-            : state.coinPrices[0] >
-              state.coinPrices[state.coinPrices.length - 1]
+            : state.coinData.coinPrices[0] >
+              state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
             ? theme.btcPriceChartBorderColorLoss
             : theme.btcPriceChartBorderColorGain,
         backgroundColor: (context) => {
           const ctx = context.chart.ctx;
           const gradient = ctx.createLinearGradient(0, 0, 0, 350);
-          state.coinPrices[0] < state.coinPrices[state.coinPrices.length - 1]
+          state.coinData.coinPrices[0] <
+          state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
             ? gradient.addColorStop(0, theme.btcPriceChartGradienColorGain)
-            : state.coinPrices[0] >
-              state.coinPrices[state.coinPrices.length - 1]
+            : state.coinData.coinPrices[0] >
+              state.coinData.coinPrices[state.coinData.coinPrices.length - 1]
             ? gradient.addColorStop(0, theme.btcPriceChartGradienColorLoss)
             : gradient.addColorStop(0, theme.btcPriceChartGradienColorGain);
           gradient.addColorStop(1, "rgba(0, 0, 0, 0.0)");
@@ -202,29 +186,37 @@ export const Coin = (props) => {
         <CoinBoxContainer>
           <CoinBox>
             <CoinBoxInner>
-              <CoinImage src={state.coinImgSrc} />
+              <CoinImage src={state.coinData.coinImgSrc} />
             </CoinBoxInner>
             <h2>
-              {state.coinName} ({state.coinSymbol})
+              {state.coinData.coinName} ({state.coinData.coinSymbol})
             </h2>
           </CoinBox>
           <CoinUrlBox>
             <UrlIcon />
-            <HomePageDiv>{state.coinHomePage}</HomePageDiv>
+            <HomePageDiv>{state.coinData.coinHomePage}</HomePageDiv>
           </CoinUrlBox>
         </CoinBoxContainer>
         <PriceBoxContainer>
           <h2 style={{ fontSize: "44px" }}>
-            {formatNumber(state.coinCurrentPrice)}
+            {formatNumber(state.coinData.coinCurrentPrice)}
           </h2>
           <PriceCHangePercentageDiv>
             <PriceChangePercentage
-              color={state.priceChangePercentage24h > 0 ? "#00fc2a" : "#fe1040"}
+              color={
+                state.coinData.priceChangePercentage24h > 0
+                  ? "#00fc2a"
+                  : "#fe1040"
+              }
             >
               {" "}
-              {state.priceChangePercentage24h}%
+              {state.coinData.priceChangePercentage24h}%
             </PriceChangePercentage>
-            {state.priceChangePercentage24h > 0 ? <ArrowUp /> : <ArrowDown />}
+            {state.coinData.priceChangePercentage24h > 0 ? (
+              <ArrowUp />
+            ) : (
+              <ArrowDown />
+            )}
           </PriceCHangePercentageDiv>
           <SquareStackIcon />
           <AllTimeContainer>
@@ -233,9 +225,10 @@ export const Coin = (props) => {
             </ArrowSvgDiv>
             <AllTimeDiv>
               <AllTimeP>
-                All Time High: <span> {formatNumber(state.coinAth)}</span>
+                All Time High:{" "}
+                <span> {formatNumber(state.coinData.coinAth)}</span>
               </AllTimeP>
-              <p>{state.coinAthDate}</p>
+              <p>{state.coinData.coinAthDate}</p>
             </AllTimeDiv>
           </AllTimeContainer>
           <AllTimeContainer>
@@ -244,9 +237,10 @@ export const Coin = (props) => {
             </ArrowSvgDiv>
             <AllTimeDiv>
               <AllTimeP>
-                All Time Low: <span> {formatNumber(state.coinAtl)}</span>
+                All Time Low:{" "}
+                <span> {formatNumber(state.coinData.coinAtl)}</span>
               </AllTimeP>
-              <p>{state.coinAtlDate}</p>
+              <p>{state.coinData.coinAtlDate}</p>
             </AllTimeDiv>
           </AllTimeContainer>
         </PriceBoxContainer>
@@ -256,7 +250,8 @@ export const Coin = (props) => {
               <PlusSmallSvg />
             </SmallRectangleBlue>
             <p>
-              Market Cap: <span>{formatNumber(state.coinMarketCap)}</span>{" "}
+              Market Cap:{" "}
+              <span>{formatNumber(state.coinData.coinMarketCap)}</span>{" "}
             </p>
           </CoinDataRow>
           <CoinDataRow>
@@ -265,7 +260,9 @@ export const Coin = (props) => {
             </SmallRectangleBlue>
             <p>
               Fully Diluted Valuation:{" "}
-              <span>{formatNumber(state.coinFullyDillutedValuation)}</span>
+              <span>
+                {formatNumber(state.coinData.coinFullyDillutedValuation)}
+              </span>
             </p>
           </CoinDataRow>
           <CoinDataRow>
@@ -274,7 +271,7 @@ export const Coin = (props) => {
             </SmallRectangleBlue>
             <p>
               Trading Volume 24h:{" "}
-              <span>{formatNumber(state.coinVolume24h)}</span>
+              <span>{formatNumber(state.coinData.coinVolume24h)}</span>
             </p>
           </CoinDataRow>
           <CoinDataRow>
@@ -282,7 +279,8 @@ export const Coin = (props) => {
               <PlusSmallSvg />
             </SmallRectangleBlue>
             <p>
-              Volume / Market: <span>{state.coinVolumeOverMarketCap}%</span>
+              Volume / Market:{" "}
+              <span>{state.coinData.coinVolumeOverMarketCap}%</span>
             </p>
           </CoinDataRow>
           <CoinDataRow>
@@ -292,7 +290,8 @@ export const Coin = (props) => {
             <p>
               Circulating Supply:{" "}
               <span>
-                {state.coinCirculatingSupply} {state.coinSymbol}
+                {state.coinData.coinCirculatingSupply}{" "}
+                {state.coinData.coinSymbol}
               </span>
             </p>
           </CoinDataRow>
@@ -303,19 +302,19 @@ export const Coin = (props) => {
             <p>
               Max Supply:{" "}
               <span>
-                {state.coinTotalSupply} {state.coinSymbol}
+                {state.coinData.coinTotalSupply} {state.coinData.coinSymbol}
               </span>
             </p>
           </CoinDataRow>
           <CoinDataRow style={{ width: "50%", marginLeft: "-180px" }}>
             <CoinsPercentageBar
-              num1={state.num1 + "%"}
-              num2={state.num2 + "%"}
-              width={state.num1 + "%"}
-              color1={state.color1}
-              color2={state.color2}
-              background1={state.color1}
-              background2={state.color2}
+              num1={state.coinData.num1 + "%"}
+              num2={state.coinData.num2 + "%"}
+              width={state.coinData.num1 + "%"}
+              color1={state.coinData.color1}
+              color2={state.coinData.color2}
+              background1={state.coinData.color1}
+              background2={state.coinData.color2}
             />
           </CoinDataRow>
         </CoinDataBoxContainer>
@@ -325,27 +324,29 @@ export const Coin = (props) => {
       </HeaderDiv>
       <DescriptionWrapper>
         <SquareStackIcon />
-        <CoinDescriptionDiv>{state.coinDescription}</CoinDescriptionDiv>
+        <CoinDescriptionDiv>
+          {state.coinData.coinDescription}
+        </CoinDescriptionDiv>
       </DescriptionWrapper>
       <CoinUrlsRow>
         <UrlDiv>
           <UrlIcon />
-          <UrlAddressDiv>{state.coinBlockChainSite1}</UrlAddressDiv>
+          <UrlAddressDiv>{state.coinData.coinBlockChainSite1}</UrlAddressDiv>
           <CopyIcon />
         </UrlDiv>
         <UrlDiv>
           <UrlIcon />
-          <UrlAddressDiv>{state.coinBlockChainSite2}</UrlAddressDiv>
+          <UrlAddressDiv>{state.coinData.coinBlockChainSite2}</UrlAddressDiv>
           <CopyIcon />
         </UrlDiv>
         <UrlDiv>
           <UrlIcon />
-          <UrlAddressDiv>{state.coinBlockChainSite3}</UrlAddressDiv>
+          <UrlAddressDiv>{state.coinData.coinBlockChainSite3}</UrlAddressDiv>
           <CopyIcon />
         </UrlDiv>
       </CoinUrlsRow>
       <CurrencyConversionRow>
-        <CurrencyConverter coinSymbol={state.coinSymbol} />
+        <CurrencyConverter coinSymbol={state.coinData.coinSymbol} />
       </CurrencyConversionRow>
       <BigChartWrapper>
         <BigLineChart data={coinPricesData} />
